refactor: avoids cyclic dependency and uses one export for all queries
