+++ web/src/hooks/graphql/utils.ts
@@ -11,6 +11,7 @@ import {
 import { useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useSnackbar } from 'notistack';
+
 import routing from '../../constants/routing';
 import { Scalars, Exact } from '../../@types/graphql';
 import { InputMutationTuple } from '../../@types/helpers';
